requirements.txt

langchain_core==0.2.34
langchain_google_genai==1.0.10
langchain_groq==0.1.9
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1